MemberInfoWeb/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.4
django-background-tasks==1.2.5
django-compat==1.0.15
django-rest-framework==0.1.0
djangorestframework==3.15.1
flake8==7.0.0
idna==3.7
isort==5.13.2
mccabe==0.7.0
pycodestyle==2.11.1
pyflakes==3.2.0
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.1
